+++ src/game/config.cpp
@@ -188,7 +188,7 @@ void map_allow_add(int index)
 	s_set_map_allows.insert(index);
 }
 
-void map_allow_copy(long * pl, int size)
+void map_allow_copy(int32_t* pl, int size)
 {
 	int iCount = 0;
 	std::set<int>::iterator it = s_set_map_allows.begin();
