+++ src/game/map_location.h
@@ -6,13 +6,13 @@ class CMapLocation : public singleton<CMapLocation>
 	public:
 		typedef struct SLocation
 		{
-			long        addr;
-			WORD        port;
+			uint32_t        addr;
+			uint16_t        port;
 		} TLocation;    
 
-		bool    Get(long x, long y, long & lMapIndex, long & lAddr, WORD & wPort);
-		bool	Get(int iIndex, long & lAddr, WORD & wPort);
-		void    Insert(long lIndex, const char * c_pszHost, WORD wPort);
+		bool    Get(long x, long y, int32_t& lMapIndex, uint32_t& lAddr, uint16_t& wPort);
+		bool	Get(int32_t iIndex, uint32_t& lAddr, uint16_t& wPort);
+		void    Insert(int32_t lIndex, const char * c_pszHost, uint16_t wPort);
 
 	protected:
 		std::map<long, TLocation> m_map_address;
